AIML_Unsupervised Learning
Bank Customer Segmentation: Identified 3 different categories of customers based on credit limit and mode of 
communication with the bank
 ● Segment 0 – High Credit Limit
 ● Preference of online banking, they are the the high spenders with multiple credit cards, and make up about 8% 
of the customer base.  They can be targeted by online campaigns and incentivized for new products.
 ● Segment 1 – Medium Credit Limit
 ● Perference of in-person banking, the make up majority of the customer base (58%).  More customers should be 
incentivized use online services for better tracking and resolution time for issues, as well as being able to be 
targeted for online campaigns.  Additionally, in-person specials should be introduced to upsell the customers. 
● Segment 2 – Low Credit Limit
 ● Preference of phone banking, make up about 34% of the customer base. Should be incentivized to use online 
services or in-person.  This will make it easier to target them for upselling and  also make it easier to track 
and improve customer servic